feat: rename component Card
